+++ pyproject.toml
@@ -30,6 +30,8 @@ ruamel-yaml = "^0.18.5"
 pytest-cov = "^4.1.0"
 arrow = "^1.3.0"
 pydash = "^7.0.6"
+pandas = "^2.2.0"
+matplotlib = "^3.8.2"
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "^3.6.0"
